Refactor document node instances deletion

+++ mayan/apps/document_indexing/managers.py
@@ -33,7 +33,10 @@ class IndexManager(models.Manager):
 
         # Only update indexes where the document type is found or that do not have any document type specified
         for index in self.get_query_set().filter(Q(enabled=True) & (Q(document_types=None) | Q(document_types=document.document_type))):
-            index.template_root.evaluate(eval_dict, document)
+            try:
+                index.template_root.evaluate(eval_dict, document)
+            except Exception as exception:
+                logger.warning('exception: %s' % exception)
 
 
 class IndexInstanceNodeManager(models.Manager):
@@ -42,7 +45,7 @@ class IndexInstanceNodeManager(models.Manager):
         Delete all the index instances related to a document
         """
         for node_instance in document.indexinstancenode_set.all():
-            index_warnings = self.cascade_document_remove(node_instance, document)
+            self.cascade_document_remove(node_instance, document)
 
     def cascade_document_remove(self, node_instance, document):
         """
@@ -50,12 +53,13 @@ class IndexInstanceNodeManager(models.Manager):
         recusively deleting documents and empty index instances up to the
         root of the tree
         """
-        remove_document_from_node_instance.send(sender=node_instance, document=document)
-        node_instance.documents.remove(document)
-        if node_instance.documents.count() == 0 and node_instance.get_children().count() == 0:
-            # if there are no more documents and no children, delete
-            # node and check parent for the same conditions
-            parent = node_instance.parent
-            delete_node_instance.send(sender=node_instance)
-            node_instance.delete()
-            self.cascade_document_remove(parent, document)
+        if not node_instance.is_root_node():
+            remove_document_from_node_instance.send(sender=node_instance, document=document)
+            node_instance.documents.remove(document)
+            if node_instance.documents.count() == 0 and node_instance.get_children().count() == 0:
+                # if there are no more documents and no children, delete
+                # node and check parent for the same conditions
+                parent = node_instance.parent
+                delete_node_instance.send(sender=node_instance)
+                node_instance.delete()
+                self.cascade_document_remove(parent, document)
